+++ jdplus-main-desktop/jdplus-toolkit-desktop-plugin/src/main/java/internal/toolkit/desktop/plugin/tsproviders/PasteProvidersNodeAction.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2013 National Bank of Belgium
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
+ * by the European Commission - subsequent versions of the EUPL (the "Licence");
+ * You may not use this work except in compliance with the Licence.
+ * You may obtain a copy of the Licence at:
+ *
+ * http://ec.europa.eu/idabc/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the Licence is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Licence for the specific language governing permissions and
+ * limitations under the Licence.
+ */
+package internal.toolkit.desktop.plugin.tsproviders;
+
+import jdplus.toolkit.desktop.plugin.TsManager;
+import jdplus.toolkit.desktop.plugin.datatransfer.DataSourceTransferManager;
+import jdplus.toolkit.desktop.plugin.datatransfer.DataTransfers;
+import nbbrd.design.ClassNameConstant;
+import org.openide.awt.ActionID;
+import org.openide.awt.ActionRegistration;
+import org.openide.nodes.Node;
+import org.openide.util.HelpCtx;
+import org.openide.util.NbBundle.Messages;
+import org.openide.util.actions.NodeAction;
+
+@ActionID(category = "Edit", id = PasteProvidersNodeAction.ID)
+@ActionRegistration(displayName = "#CTL_PasteProvidersAction", lazy = false)
+@Messages("CTL_PasteProvidersAction=Paste")
+public final class PasteProvidersNodeAction extends NodeAction {
+
+    @ClassNameConstant
+    public static final String ID = "internal.toolkit.desktop.plugin.tsproviders.PasteProvidersNodeAction";
+
+    @Override
+    protected void performAction(Node[] activatedNodes) {
+        DataSourceTransferManager.get()
+                .getDataSource(DataTransfers.systemClipboardAsTransferable())
+                .ifPresent(source -> TsManager.get().open(source));
+    }
+
+    @Override
+    protected boolean enable(Node[] activatedNodes) {
+        return DataSourceTransferManager.get().canHandle(DataTransfers.systemClipboardAsTransferable());
+    }
+
+    @Override
+    public String getName() {
+        return Bundle.CTL_PasteProvidersAction();
+    }
+
+    @Override
+    public HelpCtx getHelpCtx() {
+        return null;
+    }
+}